+++ src/game/core/progression/index.ts
@@ -2,6 +2,7 @@ export * from "./AffixRules";
 export * from "./DuplicateRewardRules";
 export * from "./EquipmentStatRules";
 export * from "./HeroStatRules";
+export * from "./HeroLevelRules";
 export * from "./ItemRewardRules";
 export * from "./LevelingRules";
 export * from "./SkillRules";